backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
anthropic
groq
pytest
httpx
